# Accelerate `bincount`, `histogram2d`, `histogramdd` with CUB