--- a/cupy/_statistics/histogram.py
+++ b/cupy/_statistics/histogram.py
@@ -545,7 +545,7 @@ def bincount(x, weights=None, minlength=None):
         if minlength < 0:
             raise ValueError('minlength must be non-negative')
 
-    size = int(cupy.max(x)) + 1
+    size = int(cupy.max(x)) + 1  # synchronize!
     if minlength is not None:
         size = max(size, minlength)