--- a/cupy/cuda/cub.pyx
+++ b/cupy/cuda/cub.pyx
@@ -3,6 +3,7 @@
 """Wrapper of CUB functions for CuPy API."""
 
 from cpython cimport sequence
+from libc.stdint cimport intptr_t
 
 from cupy_backends.cuda.api cimport runtime
 from cupy._core.core cimport _internal_ascontiguousarray
@@ -58,6 +59,8 @@ cdef extern from 'cupy_cub.h' nogil:
     void cub_device_scan(void*, size_t&, void*, void*, int, Stream_t, int, int)
     void cub_device_histogram_range(void*, size_t&, void*, void*, int, void*,
                                     size_t, Stream_t, int)
+    void cub_device_histogram_even(void*, size_t&, void*, void*, int, int, int,
+                                   size_t, Stream_t, int)
     size_t cub_device_reduce_get_workspace_size(void*, void*, int, Stream_t,
                                                 int, int)
     size_t cub_device_segmented_reduce_get_workspace_size(
@@ -68,6 +71,8 @@ cdef extern from 'cupy_cub.h' nogil:
         void*, void*, int, Stream_t, int, int)
     size_t cub_device_histogram_range_get_workspace_size(
         void*, void*, int, void*, size_t, Stream_t, int)
+    size_t cub_device_histogram_even_get_workspace_size(
+        void*, void*, int, int, int, size_t, Stream_t, int)
 
     # Build-time version
     int CUPY_CUB_VERSION_CODE
@@ -345,7 +350,7 @@ def device_scan(ndarray x, op):
     return x
 
 
-def device_histogram(ndarray x, ndarray bins, ndarray y):
+def device_histogram(ndarray x, ndarray y, bins):
     cdef memory.MemoryPointer ws
     cdef size_t ws_size, n_samples
     cdef int dtype_id, n_bins
@@ -354,28 +359,45 @@ def device_histogram(ndarray x, ndarray bins, ndarray y):
     cdef void* y_ptr
     cdef void* ws_ptr
     cdef Stream_t s
+    cdef bint is_even
 
     # TODO(leofang): perhaps not needed?
     # y is guaranteed contiguous
     x = _internal_ascontiguousarray(x)
-    bins = _internal_ascontiguousarray(bins)
+    if isinstance(bins, ndarray):
+        bins = _internal_ascontiguousarray(bins)
+        bins_ptr = <void*><intptr_t>bins.data.ptr
+        n_bins = bins.size
+        assert y.size == n_bins - 1
+        is_even = False
+    else:
+        n_bins = bins
+        is_even = True
 
     x_ptr = <void*>x.data.ptr
     y_ptr = <void*>y.data.ptr
-    n_bins = bins.size
-    bins_ptr = <void*>bins.data.ptr
     n_samples = x.size
     s = <Stream_t>stream.get_current_stream_ptr()
     dtype_id = common._get_dtype_id(x.dtype)
-    assert y.size == n_bins - 1
-    ws_size = cub_device_histogram_range_get_workspace_size(
-        x_ptr, y_ptr, n_bins, bins_ptr, n_samples, s, dtype_id)
 
+    if is_even:
+        ws_size = cub_device_histogram_even_get_workspace_size(
+            x_ptr, y_ptr, n_bins, 0, n_bins, n_samples, s, dtype_id)
+    else:
+        ws_size = cub_device_histogram_range_get_workspace_size(
+            x_ptr, y_ptr, n_bins, bins_ptr, n_samples, s, dtype_id)
     ws = memory.alloc(ws_size)
     ws_ptr = <void*>ws.ptr
+
     with nogil:
-        cub_device_histogram_range(ws_ptr, ws_size, x_ptr, y_ptr, n_bins,
-                                   bins_ptr, n_samples, s, dtype_id)
+        if is_even:
+            cub_device_histogram_even(
+                ws_ptr, ws_size, x_ptr, y_ptr, n_bins, 0, n_bins,
+                n_samples, s, dtype_id)
+        else:
+            cub_device_histogram_range(
+                ws_ptr, ws_size, x_ptr, y_ptr, n_bins,
+                bins_ptr, n_samples, s, dtype_id)
     return y