--- a/cupy/cuda/cub.pyx
+++ b/cupy/cuda/cub.pyx
@@ -372,6 +372,8 @@ def device_histogram(ndarray x, ndarray y, bins):
     else:
         n_bins = bins
         is_even = True
+        if runtime._is_hip_environment:
+            raise RuntimeError("not supported yet")
     assert y.size == n_bins - 1
 
     x_ptr = <void*>x.data.ptr