--- a/cupy/cuda/cupy_cub.h
+++ b/cupy/cuda/cupy_cub.h
@@ -30,11 +30,13 @@ void cub_device_segmented_reduce(void*, size_t&, void*, void*, int, int, cudaStr
 void cub_device_spmv(void*, size_t&, void*, void*, void*, void*, void*, int, int, int, cudaStream_t, int);
 void cub_device_scan(void*, size_t&, void*, void*, int, cudaStream_t, int, int);
 void cub_device_histogram_range(void*, size_t&, void*, void*, int, void*, size_t, cudaStream_t, int);
+void cub_device_histogram_even(void*, size_t&, void*, void*, int, int, int, size_t, cudaStream_t, int);
 size_t cub_device_reduce_get_workspace_size(void*, void*, int, cudaStream_t, int, int);
 size_t cub_device_segmented_reduce_get_workspace_size(void*, void*, int, int, cudaStream_t, int, int);
 size_t cub_device_spmv_get_workspace_size(void*, void*, void*, void*, void*, int, int, int, cudaStream_t, int);
 size_t cub_device_scan_get_workspace_size(void*, void*, int, cudaStream_t, int, int);
 size_t cub_device_histogram_range_get_workspace_size(void*, void*, int, void*, size_t, cudaStream_t, int);
+size_t cub_device_histogram_even_get_workspace_size(void*, void*, int, int, int, size_t, cudaStream_t, int);
 
 // This is for CUB's HistogramRange; hipCUB does not need this (see comment in cupy_cub.cu)
 #ifdef __CUDA_ARCH__
@@ -63,6 +65,9 @@ void cub_device_scan(...) {
 void cub_device_histogram_range(...) {
 }
 
+void cub_device_histogram_even(...) {
+}
+
 size_t cub_device_reduce_get_workspace_size(...) {
     return 0;
 }
@@ -80,6 +85,10 @@ size_t cub_device_scan_get_workspace_size(...) {
 }
 
 size_t cub_device_histogram_range_get_workspace_size(...) {
+    return 0;
+}
+
+size_t cub_device_histogram_even_get_workspace_size(...) {
     return 0;
 }