This PR accelerates all histogram related operations. The key insight is `bincount` is internally used to support multi-dimensional histograms (`histogram2d` and `histogramdd`), so if no weighting is required they can all be accelerated using CUB. 

Under the hood, we use CUB's `HistogramEven` instead of `HistogramRange`, as the former does not require constructing an extra `bins` array.

* `cupy.bincount()`:

(Here we also compare the timing of `cupy.histogram()` to show the benefit of not constructing a `bins` array, as for 1D integer inputs the result of the two APIs are identical.)

GPU elapsed time (us) | without CUB | with CUB | speedup | histogram (with CUB)
-----|-----|-----|-----|-----
size=100 | 96.36800214648247  | 118.18240210413933  | 0.8154175277429666  | 198.46399873495102
size=1000 | 97.90400192141533  | 120.51840275526047  | 0.8123572805742478  | 199.99359846115112
size=10000 | 97.59360253810883  | 129.33760136365888  | 0.7545648095305605  | 215.53920060396194
size=100000 | 193.79840195178988  | 130.36480098962784  | 1.4865853396056576  | 216.89600199460983
size=1000000 | 1473.8304138183594  | 152.9728040099144  | 9.634591085372524  | 252.87359803915024
size=10000000 | 16550.16632080078  | 510.9504044055939  | 32.39094475334481  | 672.3680019378662
size=100000000 | 160673.51531982422  | 4273.424053192139  | 37.59830836347851  | 4860.195207595825
size=1000000000 | 1623228.3813476562  | 42328.556823730476  | 38.34830438720823  | 46743.06869506836

* `cupy.histogram2d()`:

 GPU elapsed time (us) | without CUB | with CUB | speedup
-----|-----|-----|-----
shape=(100, 2) | 934.396803379059  | 913.814401626587  | 1.0225236127990929
shape=(1000, 2) | 937.7823948860167  | 976.0864019393922  | 0.9607575651322783
shape=(10000, 2) | 991.2000000476837  | 1032.803201675415  | 0.9597181713222398
shape=(100000, 2) | 1177.3823976516726  | 1157.2927951812744  | 1.0173591355221834
shape=(1000000, 2) | 5112.409543991089  | 3885.055994987488  | 1.3159165660899448
shape=(10000000, 2) | 46488.457107543945  | 31364.92462158204  | 1.482179781026972
shape=(100000000, 2) | 459340.2465820312  | 308182.36083984375  | 1.4904819514337526

* `cupy.histogramdd()`:

(Using a 3D histogram as an example)

 GPU elapsed time (us) | without CUB | with CUB | speedup
-----|-----|-----|-----
shape=(100, 3) | 1193.2639956474304  | 1229.7056078910828  | 0.9703655801764222
shape=(1000, 3) | 1294.3391919136045  | 1331.2928080558777  | 0.9722423076886915
shape=(10000, 3) | 1292.953610420227  | 1332.8224062919617  | 0.9700869405529778
shape=(100000, 3) | 1613.651192188263  | 1601.974391937256  | 1.0072890055607482
shape=(1000000, 3) | 6862.847948074342  | 5752.640056610107  | 1.1929910233456278
shape=(10000000, 3) | 61017.379379272454  | 48613.85307312012  | 1.2551438637767753
shape=(100000000, 3) | 612207.4829101562  | 488465.4846191406  | 1.2533280286681832

* Benchmark script:
<details><summary>Click to unfold</summary>
<p>

```python
import cupy as cp
from cupyx.profiler import benchmark

rand_max = 30
n_warmup = 3
n_repeat = 10

print(' GPU elapsed time (us) | without CUB | with CUB | speedup | histogram (with CUB)')
print('-----|-----|-----|-----|-----')
for size in range(2, 10):
    size = 10**size
    a = cp.random.randint(rand_max, size=size)
    print(f"size={size} | ", end='')

    cp._core.set_routine_accelerators([])
    cp._core.set_reduction_accelerators([])
    result = benchmark(cp.bincount, (a,), n_warmup=n_warmup, n_repeat=n_repeat)
    result_old = cp.average(result.gpu_times[0])
    print(10**6 * result_old, ' | ', end='')

    cp._core.set_routine_accelerators(['cub'])
    cp._core.set_reduction_accelerators(['cub'])
    result = benchmark(cp.bincount, (a,), n_warmup=n_warmup, n_repeat=n_repeat)
    result_new = cp.average(result.gpu_times[0])
    print(10**6 * result_new, ' | ', end='')

    print(result_old/result_new, ' | ', end='')

    result = benchmark(cp.histogram, (a,), {'bins': rand_max}, n_warmup=n_warmup, n_repeat=n_repeat)
    print(10**6 * cp.average(result.gpu_times[0]))

    out1 = cp.bincount(a)
    out2, _ = cp.histogram(a, bins=rand_max)
    assert (out1 == out2).all(), f"out1={out1}, out2={out2}"

print(' GPU elapsed time (us) | without CUB | with CUB | speedup')
print('-----|-----|-----|-----')
for size in range(2, 9):
    size = 10**size
    x = cp.random.randint(rand_max, size=size)
    y = cp.random.randint(rand_max, size=size)
    print(f"shape={(size, 2)} | ", end='')

    cp._core.set_routine_accelerators([])
    cp._core.set_reduction_accelerators([])
    result = benchmark(cp.histogram2d, (x, y), n_warmup=n_warmup, n_repeat=n_repeat)
    result_old = cp.average(result.gpu_times[0])
    print(10**6 * result_old, ' | ', end='')

    cp._core.set_routine_accelerators(['cub'])
    cp._core.set_reduction_accelerators(['cub'])
    result = benchmark(cp.histogram2d, (x, y), n_warmup=n_warmup, n_repeat=n_repeat)
    result_new = cp.average(result.gpu_times[0])
    print(10**6 * result_new, ' | ', end='')

    print(result_old/result_new)

print(' GPU elapsed time (us) | without CUB | with CUB | speedup')
print('-----|-----|-----|-----')
for size in range(2, 10):
    size = 10**size
    x = cp.random.randint(rand_max, size=3*size).reshape(size, 3)
    print(f"shape={(size, 3)} | ", end='')

    cp._core.set_routine_accelerators([])
    cp._core.set_reduction_accelerators([])
    result = benchmark(cp.histogramdd, (x,), n_warmup=n_warmup, n_repeat=n_repeat)
    result_old = cp.average(result.gpu_times[0])
    print(10**6 * result_old, ' | ', end='')

    cp._core.set_routine_accelerators(['cub'])
    cp._core.set_reduction_accelerators(['cub'])
    result = benchmark(cp.histogramdd, (x,), n_warmup=n_warmup, n_repeat=n_repeat)
    result_new = cp.average(result.gpu_times[0])
    print(10**6 * result_new, ' | ', end='')

    print(result_old/result_new)
```

</p>
</details>